% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_data.R
\name{read_abs_api}
\alias{read_abs_api}
\title{ABS API interface}
\usage{
read_abs_api(
  dataset,
  max_date = Sys.Date(),
  min_date = max_date - 180,
  ...,
  max_query_size = 10000L,
  cl = NULL,
  api_key = NULL,
  detail = "full",
  method = "libcurl",
  mode = "w"
)
}
\arguments{
\item{dataset}{Single length character string containing dataset name}

\item{max_date}{Single length date for starting request date}

\item{min_date}{Single length date for end request date}

\item{...}{Optional vectors of required parameter values (e.g.
\code{"State of Departure" = c("New South Wales", "Victoria")})}

\item{max_query_size}{Single length integer of maximum number of observations
any one API call can request. Automatically splits requests into multiple
calls.}

\item{cl}{A cluster object created by \code{parallel::makeCluster()}, or an
integer to indicate number of child-processes (integer values are ignored on
Windows) for parallel evaluations}

\item{api_key}{Unused}

\item{detail}{The desired amount of detail to be returned as per API documentation}

\item{method}{Method to be used for downloading files. Current download
methods are "internal", "wininet" (Windows only) "libcurl", "wget" and
"curl", and there is a value "auto": see ‘Details’ and ‘Note’ in
\code{download.file()}. The method  can also be set through the option
"download.file.method": see \code{options()}.}

\item{mode}{character. The mode with which to write the file. Useful values
are "w", "wb" (binary), "a" (append) and "ab". Not used for methods
"wget" and "curl".}
}
\value{
a data frame of requested data with both dimension codes and names.
}
\description{
ABS API interface
}
\examples{
df <- read_abs_api(
  "MERCH_EXP",
  "Merchandise Country" = "India",
  CL_MERCH_SITC = c("01", "05")
  )

}
