% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_meta.R
\name{read_abs_meta}
\alias{read_abs_meta}
\title{Get ABS metadata}
\usage{
read_abs_meta(
  dataset,
  structure_type = "dataflow",
  references = "all",
  detail = "referencepartial",
  method = "libcurl",
  mode = "w"
)
}
\arguments{
\item{dataset}{Single length character string containing dataset name,
metadata URL or downloaded metadata xml file}

\item{structure_type}{Data structure. One of datastructure, dataflow,
codelist, conceptscheme, categoryscheme, contentconstraint, actualconstraint,
agencyscheme, categorisation, hierarchicalcodelist}

\item{references}{Determines which artefacts referenced by the artefact are
returned. One of none, parents, parentsandsiblings, children, descendants,
all.}

\item{detail}{Desired amount of detail to be returned. One of allstubs,
referencestubs, referencepartial, allcompletestubs, referencecompletestubs,
full.}

\item{method}{Method to be used for downloading files. Current download
methods are "internal", "wininet" (Windows only) "libcurl", "wget" and
"curl", and there is a value "auto": see ‘Details’ and ‘Note’ in
\code{download.file()}. The method  can also be set through the option
"download.file.method": see \code{options()}.}

\item{mode}{character. The mode with which to write the file. Useful values
are "w", "wb" (binary), "a" (append) and "ab". Not used for methods
"wget" and "curl".}
}
\value{
A dataframe of requested ABS metadata
}
\description{
Get ABS metadata
}
\examples{
meta <- read_abs_meta("MERCH_EXP")

}
